requirements.txt
Flask
gunicorn 
Flask-sqlalchemy